initramfs: generate dependency list and cpio at the same time
Currently, this script is run twice, for the dependency list, and then
for the cpio archive.

The first one is re-run every time although its build log is suppressed
so nobody notices it.

Make it work more efficiently by generating the cpio and the dependency
list at the same time.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: Ia426288fcb9081eea801b1d1fed66da4418feaf9
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>